VoIP audio packets that are transmitted in the same tick are now merged together using Opus's repacketizer function. This helps optimize the network traffic when sending commands to or from the server.

+++ src/cl_commands.cpp
@@ -906,6 +906,6 @@ void CLIENTCOMMANDS_VoIPAudioPacket( const unsigned int frame, const unsigned ch
 
 	CLIENT_GetLocalBuffer( )->ByteStream.WriteByte( CLC_VOIPAUDIOPACKET );
 	CLIENT_GetLocalBuffer( )->ByteStream.WriteLong( frame );
-	CLIENT_GetLocalBuffer( )->ByteStream.WriteByte( length );
+	CLIENT_GetLocalBuffer( )->ByteStream.WriteShort( length );
 	CLIENT_GetLocalBuffer( )->ByteStream.WriteBuffer( data, length );
 }
